webclient: remove incorrect validation of --spam argument
This commit fixes compilation with clang, which generates the warning

error: implicit conversion from 'long' to 'double' changes value from
9223372036854775807 to 9223372036854775808
[-Werror,-Wimplicit-int-float-conversion]

for the conversion of INT64_MAX to double. Implementing a proper test
to verify that 'interval * ASYNC_S' does not overflow and can be
represented as an int64_t is not worth the effort.